refactor(mcp): split tools.ts into focused modules, tighten FTA cap to 54
tools.ts (FTA 61.26, the package's complexity hotspot and the binding file
under the --score-cap 63 gate) mixed four concerns. Split it, pure structural
move — no runbook logic added or relocated; the core trust-boundary guards
stay in registerRundownTools and mcp remains argv-build + spawn + render:

- tool-types.ts        pure type/interface decls (zero-runtime)
- tool-definitions.ts  zod input schemas + RUNDOWN_TOOL_DEFINITIONS/_NAMES
- command-builder.ts   buildRundownCommand + push* helpers
- response.ts          createMcpTextResponse / formatToolError
- tools.ts             registerRundownTools + barrel re-export (stable entry)

FTA scores per file, so the worst file drops 61.26 -> 52.99
(tool-definitions.ts). With the hotspot retired, tighten the gate from
--score-cap 63 to 54 (worst 52.99 + ~1 buffer, matching the original cap
philosophy). Also folds the repeated --input/--input-json/--input-file
triplet into a private pushRepeatableInputs helper.

Public surface is unchanged: the tools.ts barrel re-exports every previously
exported symbol, so index.ts and both test files import from ../src/tools.js
untouched. All 76 mcp tests pass with no assertion changes; tsc/biome/eslint
clean.
